+++ src/main/kotlin/com/scrumdapp/gateway/services/DownstreamRequestService.kt
@@ -33,7 +33,7 @@ class DownstreamRequestService(
 
     private var jwtToken = genGatewayToken()
 
-    fun getPassport(userId: Int): PassportContent {
+    fun getPassport(userId: Long): PassportContent {
         return performRequest(HttpMethod.GET, serviceProperties.getUrl("groups"), "/users/$userId/passport", PassportContent::class.java)
     }
